pyproject.toml
[project]
name = "labelr"
version = "0.8.0"
description = "A command-line tool to manage labeling tasks with Label Studio."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.2.0",
    "imagehash>=4.3.1",
    "label-studio-sdk>=1.0.8",
    "more-itertools>=10.5.0",
    "openfoodfacts>=2.9.0",
    "typer>=0.15.1",
    "google-cloud-batch==0.18.0",
    "huggingface-hub",
    "deepdiff>=8.6.1",
    "rapidfuzz>=3.14.3",
    "aiohttp",
    "aiofiles",
    "orjson",
]

[project.scripts]
labelr = "labelr.main:app"

[project.optional-dependencies]
ultralytics = [
    "ultralytics==8.3.223",
]
fiftyone = [
    "fiftyone~=1.10.0"
]
google = ["google-genai >= 1.56.0", "gcloud-aio-storage", "google-cloud-storage"]

[tool.uv]
package = true
